+++ tricount3.0/user.cs
@@ -11,13 +11,13 @@ namespace Tricount
         public int id { get; set; }
 
         public string nom { get; set; }
-        public int depenses { get; set; }
+        public float depenses { get; set; }
 
         public int id_soiree { get; set; }
 
         public float? dettes { get; set; }
 
-        public user(string Nom, int Depenses, int ID_Soiree, float? Dettes)
+        public user(string Nom, float Depenses, int ID_Soiree, float? Dettes)
         {
             nom = Nom;
             depenses = Depenses;
@@ -25,7 +25,7 @@ namespace Tricount
             dettes = Dettes;
         }
 
-        public user(int ID, string Nom, int Depenses, int ID_Soiree, float? Dettes)
+        public user(int ID, string Nom, float Depenses, int ID_Soiree, float? Dettes)
             : this(Nom, Depenses, ID_Soiree, Dettes)
         {
             id = ID;
